requirements.txt
Flask
gunicorn
PyPDF2
pytesseract
Pillow
pdf2image
python-docx
pandas
opencv-contrib-python-headless
numpy
ghostscript
pdf2docx
openpyxl
pdfplumber
sumy
nltk
transformers
torch
sentencepiece
protobuf
sacremoses